requirements.txt
simplejson
python-osc
discord.py
cx-Freeze